fix(ci/build-deps.sh): use CWD for `GOMODCACHE`
This fixes a packaging issue where `go-mod` was created in the root
directory instead of within the project subdirectory. As a result, the
packaging step failed when attempting to include `package/go-mod`.

By setting `GOMODCACHE` to the current directory (`$(pwd)/go-mod`),
the module cache is created in the correct location.