feat(index): extract GitHub repo slug from ebuild metadata
Add logic to extract a GitHub "owner/repo" slug from either the HOMEPAGE
or SRC_URI fields in ebuilds, if present. This is captured using a new
regular expression and returned as part of the parsed metadata.

Also refactors get_eclasses() to accept raw text instead of a file path,
avoiding the need to read the same file multiple times.

This allows the generated ebuild index to be used in `build-matrix.sh`
to match based on GitHub repository, which is more robust than matching
on package name alone. Ebuilds do not always use the same name as their
upstream projects.